sportsbuzz/requirements.txt
# Sports Buzz crawler + social ingestion
requests>=2.31.0
PyYAML>=6.0.1
python-dotenv>=1.0.0
